readers_assistant/urls.py
from django.conf import settings
from django.contrib import admin
from django.urls import path, include
from django.conf.urls.static import static
from .views import get_word_meaning


urlpatterns = [
    path('admin/', admin.site.urls),
    # path('accounts/', include('django.contrib.auth.urls')),
    path('tinymce/', include('tinymce.urls')),
    path('user/', include('user.urls')),
    path('', include('home.urls')),
    path('article/', include('article.urls')),
    path('book/', include('book.urls')),
    path('word/meaning/', get_word_meaning, name='get_word_meaning'),
]+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)

if settings.DEBUG:
    import debug_toolbar

    urlpatterns += [path("__debug__/", include(debug_toolbar.urls))]
