requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
certifi==2022.12.7
charset-normalizer==3.0.1
chirptext==0.1.2
crispy-bootstrap5==0.7
distlib==0.3.6
Django==4.1.6
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-tinymce==3.5.0
django_debug_toolbar==3.8.1
dnspython==2.3.0
fontawesomefree==6.3.0
idna==3.4
jamdict==0.1a11.post2
jamdict-data==1.5
JapaneseVerbConjugator==1.0.1
lxml==4.9.2
mongoengine==0.26.0
numpy==1.24.2
pandas==1.5.3
pandas-read-xml==0.3.1
Pillow==9.4.0
puchikarui==0.1
pyarrow==11.0.0
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
romkan==0.2.1
six==1.16.0
sqlparse==0.4.3
SudachiDict-core==20230110
SudachiPy==0.6.7
urllib3==1.26.14
xmltodict==0.13.0
zipfile36==0.1.3
